requirements.txt
absl-py==0.11.0
altair==4.1.0
argon2-cffi==20.1.0
asgiref==3.2.7
asn1crypto==1.3.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
base58==2.0.1
bleach==3.2.1
blinker==1.4
boto3==1.13.16
botocore==1.16.16
cached-property==1.5.2
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
cleo==0.8.1
click==7.1.2
clikit==0.6.0
conda==4.8.2
conda-package-handling==1.6.0
crashtest==0.3.0
cryptography==2.9.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
Django==3.0.6
docutils==0.15.2
entrypoints==0.3
enum-compat==0.0.3
flatbuffers==1.12
future==0.18.2
gast==0.3.3
gitdb==4.0.5
GitPython==3.1.11
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
importlib-metadata==3.3.0
inflect==4.1.0
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jams==0.3.4
jaraco.itertools==5.0.0
jedi==0.17.2
jeepney==0.4.3
Jinja2==2.11.2
jmespath==0.10.0
joblib==1.0.0
json5==0.9.4
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-console==6.1.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.7.0
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab==2.1.3
jupyterlab-pygments==0.1.2
jupyterlab-server==1.1.5
Keras==2.4.3
keras-compressor @ file:///home/alvarofpp/PycharmProjects/keras_compressor
Keras-Preprocessing==1.1.2
keyring==21.2.1
kiwisolver==1.2.0
llvmlite==0.34.0
lxml==4.5.1
Markdown==3.3.3
MarkupSafe==1.1.1
masonite-cli==2.2.2
matplotlib==3.2.1
MIDIUtil==1.2.1
mir-eval==0.6
mistune==0.8.4
more-itertools==8.3.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
notebook==6.0.3
numba==0.51.2
numpy==1.19.4
oauthlib==3.1.0
olefile==0.46
opencv-python==4.2.0.34
opt-einsum==3.3.0
packaging==20.8
pandas==1.0.3
pandocfilters==1.4.3
parso==0.7.1
pastel==0.2.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==5.3.0
pkginfo==1.5.0.1
prometheus-client==0.9.0
prompt-toolkit==3.0.8
protobuf==3.12.1
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser==2.20
pydeck==0.3.1
Pygments==2.7.3
pylev==1.3.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
PyQt5==5.14.2
PyQt5-sip==12.7.2
PyQtWebEngine==5.14.0
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
PyYAML==5.1.2
pyzmq==20.0.0
qtconsole==4.7.4
QtPy==1.9.0
readme-renderer==26.0
requests==2.23.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
resampy==0.2.2
rsa==4.6
ruamel-yaml==0.15.46
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
scikit-learn==0.23.2
scipy==1.5.4
SecretStorage==3.1.2
Send2Trash==1.5.0
six==1.15.0
smmap==3.0.4
sortedcontainers==2.3.0
SoundFile==0.10.3.post1
sqlparse==0.3.1
streamlit==0.60.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.1
toolz==0.10.0
tornado==6.1
tqdm==4.46.0
traitlets==5.0.5
twine==3.1.1
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.25.9
validators==0.15.0
vamp==1.1.0
watchdog==0.10.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.4.0
